third_party/blink/web_tests/external/wpt/css/css-values/calc-serialization-002-expected.txt

This is a testharness.js-based test.
[FAIL] testing calc((min(10px, 20%) + max(1rem, 2%)) * 2)
  assert_equals: expected "calc(2 * (min(10px, 20%) + max(1rem, 2%)))" but got "calc((min(10px, 20%) + max(1rem, 2%)) * 2)"
Harness: the test ran to completion.
